Remove old settings panel system
- Delete settings_screen.cpp and all panel files (wifi, firmware, time, language, display)
- Create settings_types.h for shared types (settings_wifi_network_t)
- Move helper functions (get_signal_icon, sanitize_ssid_for_display) to wifi_screen
- Migrate display_panel_init_brightness() -> display_screen_init_brightness()
- Migrate time_panel_get_24h_format() -> time_screen_get_24h_format()
- Migrate settings_screen_check_for_updates_async() -> firmware_screen_check_for_updates_async()
- Clean up settings_common.h - remove old state struct
- Update all includes to use new headers
