+++ main/CMakeLists.txt
@@ -40,19 +40,13 @@ idf_component_register(
         "api_server.cpp"
         "ota_manager.cpp"
         "auth_manager.cpp"
-        "settings/settings_screen.cpp"
         "settings/settings_menu.cpp"
         "settings/settings_wifi_screen.cpp"
-        "settings/settings_wifi_panel.cpp"
         "settings/keyboard_maps.c"
         "settings/settings_firmware_screen.cpp"
-        "settings/settings_firmware_panel.cpp"
         "settings/settings_time_screen.cpp"
-        "settings/settings_time_panel.cpp"
         "settings/settings_language_screen.cpp"
-        "settings/settings_language_panel.cpp"
         "settings/settings_display_screen.cpp"
-        "settings/settings_display_panel.cpp"
         "i18n/i18n.cpp"
         "fonts/montserrat_16_latin.c"
         "fonts/montserrat_24_latin.c"
