+++ main/settings/settings_firmware_screen.cpp
@@ -567,3 +567,85 @@ static void progress_timer_cb(lv_timer_t* timer)
             break;
     }
 }
+
+// ============================================================================
+// Background Update Check (for status bar notifications)
+// ============================================================================
+
+static firmware_update_check_cb_t s_bg_update_callback = NULL;
+static volatile bool s_bg_check_running = false;
+
+static void background_update_check_task(void* arg)
+{
+    (void)arg;
+    s_bg_check_running = true;
+    
+    bool update_available = false;
+    char latest_ver[32] = {0};
+    
+    // Allocate response buffer from PSRAM
+    char* response_buffer = (char*)heap_caps_calloc(1, HTTP_RESPONSE_BUFFER_SIZE, MALLOC_CAP_SPIRAM | MALLOC_CAP_8BIT);
+    if (response_buffer) {
+        esp_http_client_config_t config = {};
+        config.url = GITHUB_API_URL;
+        config.timeout_ms = GITHUB_API_TIMEOUT_MS;
+        config.crt_bundle_attach = esp_crt_bundle_attach;
+        config.buffer_size = 4096;
+        config.buffer_size_tx = 2048;
+        
+        esp_http_client_handle_t client = esp_http_client_init(&config);
+        if (client) {
+            esp_http_client_set_header(client, "Accept", "application/vnd.github.v3+json");
+            esp_http_client_set_header(client, "User-Agent", "ESP32-Arctic-Controller");
+            
+            esp_err_t err = esp_http_client_open(client, 0);
+            if (err == ESP_OK) {
+                int content_len = esp_http_client_fetch_headers(client);
+                int status = esp_http_client_get_status_code(client);
+                
+                if (status == 200 && content_len > 0 && content_len < HTTP_RESPONSE_BUFFER_SIZE) {
+                    int response_len = esp_http_client_read_response(client, response_buffer, HTTP_RESPONSE_BUFFER_SIZE - 1);
+                    response_buffer[response_len] = '\0';
+                    
+                    cJSON* root = cJSON_Parse(response_buffer);
+                    if (root) {
+                        cJSON* tag_name = cJSON_GetObjectItem(root, "tag_name");
+                        if (tag_name && cJSON_IsString(tag_name)) {
+                            const char* version = tag_name->valuestring;
+                            if (version[0] == 'v' || version[0] == 'V') version++;
+                            strncpy(latest_ver, version, sizeof(latest_ver) - 1);
+                            
+                            const esp_app_desc_t* app_desc = esp_app_get_description();
+                            if (app_desc && strcmp(latest_ver, app_desc->version) > 0) {
+                                update_available = true;
+                            }
+                        }
+                        cJSON_Delete(root);
+                    }
+                }
+            }
+            esp_http_client_cleanup(client);
+        }
+        free(response_buffer);
+    }
+    
+    // Call callback
+    if (s_bg_update_callback) {
+        s_bg_update_callback(update_available, latest_ver);
+        s_bg_update_callback = NULL;
+    }
+    
+    s_bg_check_running = false;
+    vTaskDelete(NULL);
+}
+
+void firmware_screen_check_for_updates_async(firmware_update_check_cb_t callback)
+{
+    if (s_bg_check_running) {
+        ESP_LOGW(TAG, "Update check already in progress, skipping");
+        return;
+    }
+    
+    s_bg_update_callback = callback;
+    xTaskCreate(background_update_check_task, "bg_fw_check", 8192, NULL, 5, NULL);
+}
